+++ internal/pkg/screens/mqtt.go
@@ -8,7 +8,7 @@ import (
 	"fyne.io/fyne/v2/widget"
 	MQTT "github.com/eclipse/paho.mqtt.golang"
 	"github.com/freedreamer82/mqtt-shell/internal/pkg/constant"
-	mqtt "github.com/freedreamer82/mqtt-shell/internal/pkg/mqtt2shell"
+	mqtt "github.com/freedreamer82/mqtt-shell/pkg/mqttchat"
 )
 
 type MqttDialog struct {
